+++ ProdManager/models/Incident.py
@@ -141,11 +141,11 @@ class Incident(db.Model):
     return result
 
   @classmethod
-  def count_by_status(cls, query, serialize=False):
+  def count_by_status(cls, query, serialize=False, filters=()):
     result = dict()
 
     for status in IncidentStatus:
       key = status.value if serialize else status
-      result[key] = ResourceHelpers.count_in_status_from_query(cls, query, status)
+      result[key] = ResourceHelpers.count_in_status_from_query(cls, query, status, filters)
 
     return result