+++ ProdManager/routes/weather/views.py
@@ -1,7 +1,7 @@
 from flask import Blueprint
 
 from ProdManager.helpers.template import custom_render_template
-
+from ProdManager.helpers.resource import resource_filters
 from ProdManager.models import (
   Monitor, Maintenance, Incident
 )
@@ -13,24 +13,27 @@ bp = Blueprint("weather", __name__)
 #############
 
 @bp.route("/monitor", methods=("GET",))
-def monitor():
-  monitors = Monitor.count_by_status(Monitor.query, serialize=True)
+@resource_filters(Monitor.filters())
+def monitor(filters):
+  monitors = Monitor.count_by_status(Monitor.query, serialize=True, filters=filters)
 
   return custom_render_template(None,
     json=dict(resources=monitors, serialize=False),
   )
 
 @bp.route("/maintenance", methods=("GET",))
-def maintenance():
-  maintenances = Maintenance.count_by_status(Maintenance.query, serialize=True)
+@resource_filters(Maintenance.filters())
+def maintenance(filters):
+  maintenances = Maintenance.count_by_status(Maintenance.query, serialize=True, filters=filters)
 
   return custom_render_template(None,
     json=dict(resources=maintenances, serialize=False),
   )
 
 @bp.route("/incident", methods=("GET",))
-def incident():
-  incidents = Incident.count_by_status(Incident.query, serialize=True)
+@resource_filters(Incident.filters())
+def incident(filters):
+  incidents = Incident.count_by_status(Incident.query, serialize=True, filters=filters)
 
   return custom_render_template(None,
     json=dict(resources=incidents, serialize=False),