+++ tests/ProdManager/models/test_models_maintenance.py
@@ -37,3 +37,9 @@ def test_count_maintenances():
 
     for status in MaintenanceStatus:
       assert type(maintenance_count[status.value]) == int
+
+  with app.app_context():
+    maintenance_count = Maintenance.count_by_status(Maintenance.query, serialize=True, filters=(Maintenance.scope_id == 1,))
+
+    for status in MaintenanceStatus:
+      assert type(maintenance_count[status.value]) == int