+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - feat(integration.Jenkins): Add Jenkins integration for updating monitors (#116) (!68)
 - feat(template.Base): Add a toggle switch to enable auto reload the page (#118) (!70)
 - feat(api.Weather): Add endpoints that return an overall count of Monitors/Incidents/Maintenances grouped by status (!71)
+  - feat(api.Weather): Allow to filter by resource fields (#120) (!72)
 - feat(api): Bump API schema to `0.2.0` (!71)
 
 ### Changed